Quick refactoring of the statistics generation
