+++ TypeCobol/LanguageServices/CodeAnalysis/Statistics/LanguageModel.Copy.txt
@@ -1,0 +1,391 @@
+InvalidToken	99,41
+	IntegerLiteral	20,27
+		UserDefinedWord	90,14
+		FILLER	5,8
+	PeriodSeparator	19,36
+		IntegerLiteral	100
+	UserDefinedWord	19,24
+		PIC	60,79
+		VALUE	19,86
+		PeriodSeparator	13,62
+	PictureCharacterString	12,85
+		PeriodSeparator	80,95
+		VALUE	8,83
+		COMP_3	8,52
+	PIC	12,85
+		PictureCharacterString	100
+	AlphanumericLiteral	5,88
+		PeriodSeparator	78,81
+		AlphanumericLiteral	20,52
+	VALUE	5,19
+		AlphanumericLiteral	89,55
+		IntegerLiteral	7,51
+	FILLER	1,18
+		PIC	92,67
+		REDEFINES	5,6
+	COMP_3	1,16
+		PeriodSeparator	84,7
+		VALUE	15,23
+	REDEFINES	0,91
+		UserDefinedWord	99,98
+	OCCURS	0,26
+		IntegerLiteral	100
+	InvalidToken	0,18
+		IntegerLiteral	98,25
+	COMP	0,11
+		PeriodSeparator	61,79
+		VALUE	37,67
+	SPACE	0,1
+		PeriodSeparator	89,84
+		LOW_VALUE	6,13
+	USAGE	0,08
+		COMP_3	87,3
+		COMP	7,62
+	THRU	0,05
+		IntegerLiteral	62,02
+		AlphanumericLiteral	37,85
+	INDEXED	0,05
+		BY	97,47
+	BY	0,05
+		UserDefinedWord	99,73
+	PartialCobolWord	0,03
+		PIC	68,35
+		VALUE	16,03
+		PeriodSeparator	10,97
+	POINTER	0,02
+		PeriodSeparator	84,3
+		VALUE	15,7
+	ZERO	0,02
+		PeriodSeparator	98,96
+	HexadecimalAlphanumericLiteral	0,02
+		PeriodSeparator	100
+	SPACES	0,02
+		PeriodSeparator	67,7
+		LOW_VALUE	14,4
+		LOW_VALUES	13,62
+	LOW_VALUE	0,01
+		PeriodSeparator	90,67
+	BLANK	0,01
+		WHEN	85,35
+		ZERO	14,65
+	WHEN	0,01
+		ZERO	100
+	BINARY	0,01
+		PeriodSeparator	49,19
+		VALUE	49,19
+	DecimalLiteral	0,01
+		PeriodSeparator	100
+	SIGN	0,01
+		LEADING	100
+	LEADING	0,01
+		SEPARATE	100
+	SEPARATE	0,01
+		PeriodSeparator	100
+	VALUES	0,01
+		IntegerLiteral	52,87
+		AlphanumericLiteral	35,63
+		ARE	9,2
+	LOW_VALUES	0
+		PeriodSeparator	56,14
+		HIGH_VALUES	43,86
+	NULL	0
+		PeriodSeparator	100
+	COMP_2	0
+		PeriodSeparator	100
+	IS	0
+		AlphanumericLiteral	63,16
+		UserDefinedWord	18,42
+		POINTER	10,53
+		COMP_3	7,89
+	RIGHT	0
+		PeriodSeparator	100
+	HIGH_VALUES	0
+		PeriodSeparator	100
+	ZEROS	0
+		PeriodSeparator	100
+	ASCENDING	0
+		KEY	96,15
+	KEY	0
+		UserDefinedWord	72
+		IS	28
+	HIGH_VALUE	0
+		PeriodSeparator	100
+	JUST	0
+		RIGHT	100
+	ALL	0
+		AlphanumericLiteral	100
+	PICTURE	0
+		PictureCharacterString	100
+	TIMES	0
+		PeriodSeparator	50
+		INDEXED	31,25
+		PIC	12,5
+		ASCENDING	6,25
+	COMP_5	0
+		PeriodSeparator	78,57
+		VALUE	21,43
+	JUSTIFIED	0
+		RIGHT	100
+	COMP_4	0
+		VALUE	77,78
+		OCCURS	22,22
+	ARE	0
+		IntegerLiteral	87,5
+		AlphanumericLiteral	12,5
+	DEPENDING	0
+		ON	100
+	ON	0
+		UserDefinedWord	100
+	QUOTE	0
+		PeriodSeparator	100
+	PACKED_DECIMAL	0
+		PeriodSeparator	100
+	TO	0
+		IntegerLiteral	100
+	ZEROES	0
+		PeriodSeparator	100
+	SECTION	0
+		PeriodSeparator	100
+END_EXEC	0,4
+	IntegerLiteral	19,04
+		UserDefinedWord	99,28
+	PeriodSeparator	18,9
+		IntegerLiteral	100
+	UserDefinedWord	18,9
+		PIC	94,8
+		PeriodSeparator	5,2
+	PIC	17,92
+		PictureCharacterString	100
+	PictureCharacterString	17,92
+		PeriodSeparator	82,4
+		USAGE	17,6
+	USAGE	3,15
+		COMP_3	88,52
+		COMP	11,48
+	COMP_3	2,79
+		PeriodSeparator	100
+	END_EXEC	0,74
+		PeriodSeparator	100
+	COMP	0,36
+		PeriodSeparator	61,9
+		OCCURS	38,1
+	OCCURS	0,14
+		IntegerLiteral	100
+	TIMES	0,14
+		PeriodSeparator	100
+COPY	0,09
+	IntegerLiteral	21,28
+		UserDefinedWord	89,23
+		FILLER	6,15
+	PeriodSeparator	20,79
+		IntegerLiteral	98,82
+	UserDefinedWord	20,62
+		PIC	63,49
+		PeriodSeparator	21,83
+		VALUE	8,33
+	PIC	14,16
+		PictureCharacterString	100
+	PictureCharacterString	14,16
+		PeriodSeparator	97,11
+	AlphanumericLiteral	2,7
+		PeriodSeparator	63,64
+		AlphanumericLiteral	36,36
+	VALUE	1,72
+		AlphanumericLiteral	100
+	FILLER	1,31
+		PIC	68,75
+		REDEFINES	31,25
+	REDEFINES	1,15
+		UserDefinedWord	100
+	OCCURS	0,74
+		IntegerLiteral	100
+	COPY	0,49
+		UserDefinedWord	100
+	BLANK	0,25
+		WHEN	100
+	WHEN	0,25
+		ZERO	100
+	ZERO	0,25
+		PeriodSeparator	100
+	SUPPRESS	0,16
+		PeriodSeparator	100
+EXEC	0,08
+	ExecStatementText	92,82
+		ExecStatementText	96,13
+	EXEC	3,59
+		ExecTranslatorName	100
+	ExecTranslatorName	3,59
+		ExecStatementText	100
+IF	0
+	UserDefinedWord	35,71
+		GO	40
+		MOVE	15
+		PERFORM	15
+		NOT	10
+		SET	10
+		EqualOperator	5
+		EVALUATE	5
+	IF	33,93
+		NOT	52,63
+		UserDefinedWord	47,37
+	NOT	21,43
+		UserDefinedWord	83,33
+		EqualOperator	16,67
+	EqualOperator	5,36
+		SPACE	66,67
+		UserDefinedWord	33,33
+	SPACE	3,57
+		MOVE	100
+PERFORM	0
+	UserDefinedWord	40,74
+		IF	40,91
+		PeriodSeparator	18,18
+		WHEN	13,64
+		RightParenthesisSeparator	9,09
+		END_IF	9,09
+	PERFORM	31,48
+		UserDefinedWord	100
+	PeriodSeparator	7,41
+		UserDefinedWord	50
+		EXIT	50
+	WHEN	5,56
+		LeftParenthesisSeparator	33,33
+		UserDefinedWord	33,33
+		OTHER	33,33
+	LeftParenthesisSeparator	5,56
+		UserDefinedWord	66,67
+		LeftParenthesisSeparator	33,33
+	RightParenthesisSeparator	5,56
+		AND	33,33
+		RightParenthesisSeparator	33,33
+		SET	33,33
+	AND	1,85
+		LeftParenthesisSeparator	100
+	OTHER	1,85
+		SET	100
+SET	0
+	SET	24
+		UserDefinedWord	100
+	UserDefinedWord	24
+		TO	100
+	TO	24
+		TRUE	100
+	TRUE	24
+		PERFORM	66,67
+		ELSE	16,67
+		END_IF	8,33
+		WHEN	8,33
+	WHEN	2
+		OTHER	100
+	OTHER	2
+		PERFORM	100
+MOVE	0
+	UserDefinedWord	33,33
+		TO	27,27
+		IF	18,18
+		ELSE	18,18
+		END_IF	18,18
+		LeftParenthesisSeparator	9,09
+		SET	9,09
+	MOVE	21,21
+		UserDefinedWord	57,14
+		LOW_VALUE	28,57
+		WHEN_COMPILED	14,29
+	TO	21,21
+		UserDefinedWord	100
+	IntegerLiteral	6,06
+		ColonSeparator	50
+		RightParenthesisSeparator	50
+	LOW_VALUE	6,06
+		TO	100
+	WHEN_COMPILED	3,03
+		TO	100
+	LeftParenthesisSeparator	3,03
+		IntegerLiteral	100
+	ColonSeparator	3,03
+		IntegerLiteral	100
+	RightParenthesisSeparator	3,03
+		TO	100
+END_IF	0
+	END_IF	70,37
+		END_IF	26,32
+		IF	26,32
+		SET	15,79
+		PERFORM	10,53
+		EVALUATE	5,26
+		WHEN	5,26
+		END_EVALUATE	5,26
+		PeriodSeparator	5,26
+	UserDefinedWord	11,11
+		NOT	33,33
+		SET	33,33
+		PeriodSeparator	33,33
+	PeriodSeparator	7,41
+		UserDefinedWord	50
+		EXIT	50
+	WHEN	3,7
+		UserDefinedWord	100
+	NOT	3,7
+		EqualOperator	100
+	EqualOperator	3,7
+		UserDefinedWord	100
+GO	0
+	GO	50
+		UserDefinedWord	100
+	UserDefinedWord	50
+		END_IF	100
+EXIT	0
+	PeriodSeparator	36,36
+		UserDefinedWord	50
+		PERFORM	50
+	EXIT	27,27
+		PeriodSeparator	100
+	UserDefinedWord	18,18
+		SECTION	100
+	SECTION	18,18
+		PeriodSeparator	100
+EVALUATE	0
+	EVALUATE	25
+		TRUE	100
+	TRUE	25
+		WHEN	100
+	WHEN	25
+		UserDefinedWord	100
+	UserDefinedWord	25
+		SET	50
+		CONTINUE	50
+END_EVALUATE	0
+	END_EVALUATE	40
+		END_IF	50
+		PeriodSeparator	50
+	PeriodSeparator	40
+		UserDefinedWord	50
+		GOBACK	50
+	UserDefinedWord	20
+		PeriodSeparator	100
+GOBACK	0
+	PeriodSeparator	40
+		UserDefinedWord	50
+		IF	50
+	GOBACK	20
+		PeriodSeparator	100
+	UserDefinedWord	20
+		SECTION	100
+	SECTION	20
+		PeriodSeparator	100
+ELSE	0
+	ELSE	100
+		IF	50
+		PERFORM	25
+		MOVE	25
+CONTINUE	0
+	CONTINUE	33,33
+		WHEN	100
+	WHEN	33,33
+		UserDefinedWord	100
+	UserDefinedWord	33,33
+		SET	100
+--- Model performances ---
+Next word is top 1 suggestion  85,88% of the time
+Next word in top 3 suggestions 97,63% of the time
+Next word in top 5 suggestions 97,63% of the time
